neural_solution/requirements.txt
fastapi
grpcio
mpi4py
neural_compressor>=2.2
protobuf
pydantic
uvicorn[standard]
watchdog
